requirements.txt
Django==5.0.7
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
drf-spectacular==0.27.2
drf-yasg==1.21.7
psycopg2-binary==2.9.9
pytest==8.3.3
pytest-django==4.9.0
pytest-ordering==0.6
PyJWT==1.7.1